Feat/login: 로그인 / 로그아웃 / 회원가입 앱 내에서 진행되게 수정

+++ PADO/PADO/iOS/Helpers/Services/UpdateHeartData.swift
@@ -12,10 +12,13 @@ import SwiftUI
 
 class UpdateHeartData {
     
+    static let shared = UpdateHeartData()
+    
     let db = Firestore.firestore()
     
     func addHeart(documentID: String) async {
         // 햅틱 피드백 생성
+        guard !userNameID.isEmpty else { return }
         
         do {
             try await db.collection("users").document(userNameID).collection("highlight").document(documentID).setData(["documentID": documentID,
@@ -51,6 +54,8 @@ class UpdateHeartData {
     }
     
     func deleteHeart(documentID: String) async {
+        guard !userNameID.isEmpty else { return }
+        
         do {
             try await db.collection("users").document(userNameID).collection("highlight").document(documentID).delete()
